refactor : 패키지이동 및 스낵바 추가, 린트정리
