+++ app/src/main/java/com/recodream_aos/recordream/presentation/document/DocumentBottomSheetFragment.kt
@@ -10,7 +10,7 @@ import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.recodream_aos.recordream.R
 import com.recodream_aos.recordream.databinding.FragmentDocumentBottomSheetBinding
-import com.recodream_aos.recordream.util.CustomDialog
+import com.recodream_aos.recordream.util.customview.CustomDialog
 
 class DocumentBottomSheetFragment : BottomSheetDialogFragment() {
     private lateinit var binding: FragmentDocumentBottomSheetBinding
@@ -20,7 +20,7 @@ class DocumentBottomSheetFragment : BottomSheetDialogFragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ): View? {
+    ): View {
         binding = DataBindingUtil.inflate(
             inflater,
             R.layout.fragment_document_bottom_sheet,
