+++ app/src/main/java/com/recodream_aos/recordream/presentation/mypage/MypageBottomSheetFragment.kt
@@ -21,7 +21,7 @@ class MypageBottomSheetFragment : BottomSheetDialogFragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ): View? {
+    ): View {
         _binding = FragmentMypageBottomSheetBinding.inflate(inflater, container, false)
         return binding.root
     }
