+++ app/src/main/java/com/recodream_aos/recordream/util/customview/CustomBottomNavigation.kt
@@ -1,4 +1,4 @@
-package com.recodream_aos.recordream.util
+package com.recodream_aos.recordream.util.customview
 
 import android.content.Context
 import android.graphics.Canvas
