+++ app/src/main/java/com/recodream_aos/recordream/util/customview/CustomDialog.kt
@@ -1,4 +1,4 @@
-package com.recodream_aos.recordream.util
+package com.recodream_aos.recordream.util.customview
 
 import android.app.Dialog
 import android.content.Context
