+++ app/src/main/java/com/recodream_aos/recordream/util/extension/ContextExtension.kt
@@ -1,11 +1,7 @@
 package com.recodream_aos.recordream.util
 
 import android.content.Context
-import android.util.Log
 import android.widget.Toast
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 import kotlin.math.roundToInt
 
 fun Context.shortToastByString(message: String) {
@@ -18,22 +14,3 @@ fun Context.shortToastByInt(message: Int) {
 
 fun Context.dpToPixel(dp: Int): Int =
     (dp * resources.displayMetrics.density).roundToInt()
-
-fun <ResponseType> Call<ResponseType>.enqueueUtil(
-    onSuccess: (ResponseType) -> Unit,
-    onError: ((stateCode: Int) -> Unit)? = null,
-) {
-    this.enqueue(object : Callback<ResponseType> {
-        override fun onResponse(call: Call<ResponseType>, response: Response<ResponseType>) {
-            if (response.isSuccessful) {
-                onSuccess.invoke(response.body() ?: return)
-            } else {
-                onError?.invoke(response.code())
-            }
-        }
-
-        override fun onFailure(call: Call<ResponseType>, t: Throwable) {
-            Log.d("NetworkTest", "error:$t")
-        }
-    })
-}
