+++ app/src/main/java/com/recodream_aos/recordream/util/extension/ViewExtension.kt
@@ -1,0 +1,17 @@
+package com.recodream_aos.recordream.util
+
+import android.graphics.Color
+import android.view.View
+import com.google.android.material.snackbar.Snackbar
+
+fun View.makeSnackBar(message: Int) {
+    Snackbar.make(this, message, Snackbar.LENGTH_SHORT).show()
+}
+
+fun View.anchorSnackBar(message: Int) {
+    Snackbar.make(this, message, Snackbar.LENGTH_SHORT).apply {
+        anchorView = this@anchorSnackBar
+        this.view.setBackgroundColor(Color.parseColor("#373737"))
+        setTextColor(Color.parseColor("#FFFFFF"))
+    }.show()
+}
